M4ATX: fixed issues with usb lookup
Previously version of the code for finding the M4ATX USB device
was always just picking last USB device in the registry.

Fixed two issues that lead to above behavior:

Make sure that comparing ProductID actually works. It's a bit unclear
what is going on, but masking out higher bits comparison works.

Checking that MyUsbDevice is null to detect that device is not found
does not work. If will be set to last device we from registry we looked
at. Use a special boolean flag instad.